# Conflicts:
#	src/main/java/it/globus/finance/rest/controller/TransactionController.java
#	src/main/java/it/globus/finance/service/TransactionService.java

+++ src/main/java/it/globus/finance/model/repo/CategoryRepo.java
@@ -3,7 +3,5 @@ package it.globus.finance.model.repo;
 import it.globus.finance.model.entity.Category;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigInteger;
-
-public interface CategoryRepo extends CrudRepository<Category, BigInteger> {
+public interface CategoryRepo extends CrudRepository<Category, Long> {
 }
