+++ src/main/java/it/globus/finance/model/repo/ReportRepo.java
@@ -3,7 +3,5 @@ package it.globus.finance.model.repo;
 import it.globus.finance.model.entity.Report;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigInteger;
-
-public interface ReportRepo extends CrudRepository<Report, BigInteger> {
+public interface ReportRepo extends CrudRepository<Report, Long> {
 }
